+++ app/src/parser/transactions/base_tx_fields.rs
@@ -174,7 +174,7 @@ where
             .get_obj_if(filter)
             .ok_or(ParserError::DisplayIdxOutOfRange)?;
 
-        Ok((obj, obj_item_n as u8))
+        Ok((obj, obj_item_n))
     }
 }
